# openapi.yaml
openapi: 3.1.0
info:
  title: AGP Control Plane API
  version: 0.2.0
  description: >
    OpenAPI contract derived from the authoritative AGP control-plane specs.
    This document defines the network contract for orchestration clients and runtimes.
    V1 authentication uses bearer tokens only; /health is intentionally unauthenticated.
servers:
  - url: http://127.0.0.1:7860
...
security:
  - bearerAuth: []
paths:
...
  /runtimes/register:
    post:
      tags: [runtimes]
      summary: Register or refresh runtime identity
      operationId: registerRuntime
      deprecated: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRuntimeRequest'
      responses:
        '200':
          description: Runtime registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeEnvelope'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /runtimes:
    get:
      tags: [runtimes]
      summary: List runtimes
      operationId: listRuntimes
      deprecated: true
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/RuntimeStatus'
        - in: query
          name: health_status
          schema:
            $ref: '#/components/schemas/HealthStatus'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Runtime list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /runs/claim:
    post:
      tags: [runs]
      summary: Claim a job pull-style
      operationId: claimRun
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimRunRequest'
      responses:
        '200':
          description: Job claimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimRunEnvelope'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /runs/{run_id}/heartbeat:
    post:
      tags: [runs]
      summary: Renew lease by heartbeat
      operationId: heartbeatRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaseActionRequest'
      responses:
        '200':
          description: Heartbeat accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaseEnvelope'
        '409':
          $ref: '#/components/responses/LeaseExpired'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /runs/{run_id}/progress:
    post:
      tags: [runs]
      summary: Publish progress
      operationId: progressRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressRequest'
      responses:
        '200':
          description: Progress accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunEnvelope'
        '409':
          $ref: '#/components/responses/StaleFencingToken'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /runs/{run_id}/recovering:
    post:
      tags: [runs]
      summary: Enter bounded local recovery
      operationId: recoveringRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryRequest'
      responses:
        '200':
          description: Recovery state accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunEnvelope'
        '409':
          $ref: '#/components/responses/StaleFencingToken'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /runs/{run_id}/resumed:
    post:
      tags: [runs]
      summary: Resume a recovering run
      operationId: resumedRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryRequest'
      responses:
        '200':
          description: Run resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunEnvelope'
        '409':
          $ref: '#/components/responses/StaleFencingToken'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /runs/{run_id}/cancel:
    post:
      tags: [runs]
      summary: Cancel a running run after interrupt
      operationId: cancelRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRunRequest'
      responses:
        '200':
          description: Run cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunEnvelope'
        '409':
          $ref: '#/components/responses/StaleFencingToken'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /runs/{run_id}/complete:
    post:
      tags: [runs]
      summary: Complete run
      operationId: completeRun
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteRunRequest'
      responses:
        '200':
          description: Run completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunEnvelope'
        '409':
          description: Stale or conflicting terminal report
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /runs/{run_id}/fail:
    post:
      tags: [runs]
      summary: Fail run
      operationId: failRun
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailRunRequest'
      responses:
        '200':
          description: Run failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunEnvelope'
        '409':
          description: Stale or conflicting terminal report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
components:
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema:
        type: string
    Cursor:
      in: query
      name: cursor
      schema:
        type: string
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
...
    RunId:
      in: path
      name: run_id
      required: true
      schema:
        type: string
...
  responses:
    InvalidRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthenticated:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    LeaseExpired:
      description: Lease expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    StaleFencingToken:
      description: Stale fencing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorCode:
      type: string
      enum:
        - invalid_request
        - not_found
        - conflict
        - stale_fencing_token
        - lease_expired
        - already_terminal
        - not_eligible
        - unauthenticated
        - forbidden
        - rate_limited
        - internal_error
    JobStatus:
      type: string
      enum: [accepted, queued, running, interrupt_requested, completed, failed, cancelled, blocked]
    RunStatus:
      type: string
      enum: [created, leased, running, recovering, completed, failed, abandoned, cancelled]
...
    RuntimeStatus:
      type: string
      enum: [registering, idle, busy, degraded, offline, draining]
    HealthStatus:
      type: string
      enum: [healthy, degraded, unreachable, draining]
    ArtifactKind:
      type: string
      enum: [prompt, transcript_log, exec_log, result, failure_evidence]
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            retryable:
              type: boolean
    PageInfo:
      type: object
      required: [has_more]
      properties:
        next_cursor:
          type: string
        has_more:
          type: boolean
...
    JobSummary:
      type: object
      required: [job_id, message_id, target_queue, status, retry_count, max_retries]
      properties:
        job_id:
          type: string
        message_id:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        target_agent_id:
          type: string
        target_queue:
          type: string
          description: Canonical queue key such as `agent:agt_123` or `capability:cap_python:v1`.
        latest_run_id:
          type: string
        retry_count:
          type: integer
        max_retries:
          type: integer
        result_artifact_id:
          type: string
    RunSummary:
      type: object
      required: [run_id, job_id, agent_id, runtime_id, attempt, status]
      properties:
        run_id:
          type: string
        job_id:
          type: string
        status:
          $ref: '#/components/schemas/RunStatus'
        attempt:
          type: integer
        agent_id:
          type: string
        runtime_id:
          type: string
...
    RuntimeSummary:
      type: object
      required: [runtime_id, hostname, release_version, status, health_status, last_heartbeat_at]
      properties:
        runtime_id:
          type: string
        hostname:
          type: string
        release_version:
          type: string
        status:
          $ref: '#/components/schemas/RuntimeStatus'
        health_status:
          $ref: '#/components/schemas/HealthStatus'
        last_heartbeat_at:
          type: string
          format: date-time
        metadata_json:
          type: object
          additionalProperties: true
...
    LeaseSummary:
      type: object
      required: [lease_id, run_id, agent_id, runtime_id, fencing_token, status, expires_at, created_at]
      properties:
        lease_id:
          type: string
        run_id:
          type: string
        agent_id:
          type: string
        runtime_id:
          type: string
        fencing_token:
          type: integer
          format: int64
        status:
          type: string
          enum: [active, expired, released]
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        released_at:
          type: string
          format: date-time
...
    RegisterRuntimeRequest:
      type: object
      required: [hostname, release_version]
      properties:
        runtime_id:
          type: string
        hostname:
          type: string
        release_version:
          type: string
        metadata:
          type: object
          additionalProperties: true
    RuntimeEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/RuntimeSummary'
    RuntimeListEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          required: [items, page]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/RuntimeSummary'
            page:
              $ref: '#/components/schemas/PageInfo'
...
    ClaimRunRequest:
      type: object
      required: [runtime_id]
      properties:
        runtime_id:
          type: string
        agent_id:
          type: string
        capability:
          type: string
        lease_ttl_seconds:
          type: integer
          minimum: 1
    ClaimRunEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          required: [claimed]
          properties:
            claimed:
              type: boolean
            job:
              $ref: '#/components/schemas/JobSummary'
            message:
              type: object
              properties:
                message_id:
                  type: string
                target_type:
                  type: string
                target_id:
                  type: string
                text:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
            run:
              $ref: '#/components/schemas/RunSummary'
            lease:
              $ref: '#/components/schemas/LeaseSummary'
            agent_id:
              type: string
    LeaseActionRequest:
      type: object
      required: [runtime_id, lease_id, fencing_token]
      properties:
        runtime_id:
          type: string
        lease_id:
          type: string
        fencing_token:
          type: integer
          format: int64
        extend_seconds:
          type: integer
    ProgressRequest:
      allOf:
        - $ref: '#/components/schemas/LeaseActionRequest'
        - type: object
          required: [message]
          properties:
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    RecoveryRequest:
      allOf:
        - $ref: '#/components/schemas/LeaseActionRequest'
        - type: object
          properties:
            details:
              type: object
              additionalProperties: true
    CancelRunRequest:
      allOf:
        - $ref: '#/components/schemas/LeaseActionRequest'
        - type: object
          properties:
            reason:
              type: string
    ArtifactReference:
      type: object
      required: [role, storage_ref, content_type, checksum, size_bytes]
      properties:
        role:
          $ref: '#/components/schemas/ArtifactKind'
        storage_ref:
          type: string
        content_type:
          type: string
        checksum:
          type: string
        size_bytes:
          type: integer
          format: int64
    CompleteRunRequest:
      allOf:
        - $ref: '#/components/schemas/LeaseActionRequest'
        - type: object
          required: [artifacts]
          properties:
            artifacts:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/ArtifactReference'
            summary:
              type: object
              additionalProperties: true
    FailRunRequest:
      allOf:
        - $ref: '#/components/schemas/LeaseActionRequest'
        - type: object
          required: [error, artifacts]
          properties:
            error:
              type: string
            artifacts:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/ArtifactReference'
            summary:
              type: object
              additionalProperties: true
    RunEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          properties:
            run_id:
              type: string
            job_id:
              type: string
            status:
              $ref: '#/components/schemas/RunStatus'
            result_artifact_id:
              type: string
    LeaseEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          properties:
            run_id:
              type: string
            lease_id:
              type: string
            status:
              $ref: '#/components/schemas/RunStatus'
            expires_at:
              type: string
              format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque